Various fixes. Test set.

+++ src/main/java/org/verapdf/io/Reader.java
@@ -34,7 +34,7 @@ public class Reader extends XRefReader {
 	}
 
 	public COSObject getObject(final long offset) throws IOException {
-		return this.parser.getObject((int) offset);
+		return this.parser.getObject(offset);
 	}
